+++ packages/wallet/shared/src/types/card.ts
@@ -1,0 +1,7 @@
+export type LockReasonCode =
+  | 'ClientRequestedLock'
+  | 'LostCard'
+  | 'StolenCard'
+  | 'IssuerRequestGeneral'
+  | 'IssuerRequestFraud'
+  | 'IssuerRequestLegal'
