+++ packages/wallet/shared/src/types/index.ts
@@ -8,3 +8,4 @@ export * from './grant'
 export * from './walletAddress'
 export * from './user'
 export * from './iframe'
+export * from './card'
